LayoutTests/fast/profiler/profile-calls-in-included-file.html

<html>
<head>
<script src="resources/profiler-test-JS-resources.js"></script>
<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();

function testStart()
{
    startProfile("Profile call in included file");
    endTest();
}
</script>
</head>

<body onload="testStart()">
This page calls the profiler from the included file.
<br>
<br>
To run this test manually, load it in the browser then load the WebInspector and look at
the profile.  If there is a profile that is sufficient for this test.
<div id="output"></div>
<script>
</script>
</body>
</html>
